Creating walls

+++ Node.h
@@ -9,9 +9,9 @@
 class Node
 {
 public:
-	Node(int x, int y);
-	int xPos;
-	int yPos;
+	Node(short x, short y);
+	short xPos;
+	short yPos;
 	float g;	// START TO NODE
 	float h;	// NODE TO GOAL
 	float f;	// F = G + H
@@ -22,9 +22,9 @@ public:
 	Node* getParent();
 	void setChild(Node* p);
 	Node* getChild();
-	float calculateFunc(int xStart, int yStart, int xFinish, int yFinish);
-	float calculateG(int xStart, int yStart);
-	float calculateH(int xFinish, int yFinish);
+	float calculateFunc(short xStart, short yStart, short xFinish, short yFinish);
+	float calculateG(short xStart, short yStart);
+	float calculateH(short xFinish, short yFinish);
 	
 private:
 	Node* parent;
